Generalize CLI structs (#3123)
Generalize so that we can use `PgoType` and adjacent CLI functions in
more use cases in a generic way.
