+++ autoprecompiles/Cargo.toml
@@ -20,6 +20,7 @@ tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.17", features = ["std", "env-filter"] }
 serde_json = "1.0.140"
 rayon = "1.10.0"
+strum = "0.27.1"
 
 [package.metadata.cargo-udeps.ignore]
 development = ["env_logger"]
