+++ autoprecompiles/src/blocks/mod.rs
@@ -9,8 +9,8 @@ mod pgo;
 mod selection;
 
 pub use detection::collect_basic_blocks;
-pub use pgo::PgoConfig;
 pub use pgo::{generate_apcs_with_pgo, ApcCandidateJsonExport, Candidate};
+pub use pgo::{pgo_config, PgoConfig, PgoType};
 pub use selection::KnapsackItem;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
