+++ cli-openvm/Cargo.toml
@@ -15,6 +15,7 @@ bench = false         # See https://github.com/bheisler/criterion.rs/issues/458
 openvm-sdk.workspace = true
 openvm-stark-sdk.workspace = true
 openvm-stark-backend.workspace = true
+powdr-autoprecompiles.workspace = true
 
 powdr-openvm.workspace = true
