+++ cli-openvm/src/main.rs
@@ -3,9 +3,9 @@ use metrics_tracing_context::{MetricsLayer, TracingContextLayer};
 use metrics_util::{debugging::DebuggingRecorder, layers::Layer};
 use openvm_sdk::StdIn;
 use openvm_stark_sdk::bench::serialize_metric_snapshot;
+use powdr_autoprecompiles::{pgo_config, PgoType};
 use powdr_openvm::{
-    default_powdr_openvm_config, CompiledProgram, GuestOptions, PgoConfig, PgoType,
-    PrecompileImplementation,
+    default_powdr_openvm_config, CompiledProgram, GuestOptions, PrecompileImplementation,
 };
 
 use clap::{CommandFactory, Parser, Subcommand};
@@ -151,14 +151,13 @@ fn run_command(command: Commands) {
             if let Some(apc_candidates_dir) = apc_candidates_dir {
                 powdr_config = powdr_config.with_apc_candidates_dir(apc_candidates_dir);
             }
-            let pgo_config = pgo_config(
-                guest.clone(),
+            let execution_profile = powdr_openvm::execution_profile_from_guest(
+                &guest,
                 guest_opts.clone(),
-                pgo,
-                max_columns,
-                max_block_instructions,
-                input,
+                stdin_from(input),
             );
+            let pgo_config =
+                pgo_config(pgo, max_columns, max_block_instructions, execution_profile);
             let program = powdr_openvm::compile_guest(
                 &guest,
                 guest_opts,
@@ -184,14 +183,13 @@ fn run_command(command: Commands) {
             if let Some(apc_candidates_dir) = apc_candidates_dir {
                 powdr_config = powdr_config.with_apc_candidates_dir(apc_candidates_dir);
             }
-            let pgo_config = pgo_config(
-                guest.clone(),
+            let execution_profile = powdr_openvm::execution_profile_from_guest(
+                &guest,
                 guest_opts.clone(),
-                pgo,
-                max_columns,
-                max_block_instructions,
-                input,
+                stdin_from(input),
             );
+            let pgo_config =
+                pgo_config(pgo, max_columns, max_block_instructions, execution_profile);
             let program = powdr_openvm::compile_guest(
                 &guest,
                 guest_opts,
@@ -220,14 +218,13 @@ fn run_command(command: Commands) {
             if let Some(apc_candidates_dir) = apc_candidates_dir {
                 powdr_config = powdr_config.with_apc_candidates_dir(apc_candidates_dir);
             }
-            let pgo_config = pgo_config(
-                guest.clone(),
+            let execution_profile = powdr_openvm::execution_profile_from_guest(
+                &guest,
                 guest_opts.clone(),
-                pgo,
-                max_columns,
-                max_block_instructions,
-                input,
+                stdin_from(input),
             );
+            let pgo_config =
+                pgo_config(pgo, max_columns, max_block_instructions, execution_profile);
             let program = powdr_openvm::compile_guest(
                 &guest,
                 guest_opts,
@@ -264,39 +261,6 @@ fn stdin_from(input: Option<u32>) -> StdIn {
         s.write(&i)
     }
     s
-}
-
-fn pgo_config(
-    guest: String,
-    guest_opts: GuestOptions,
-    pgo: PgoType,
-    max_columns: Option<usize>,
-    max_block_instructions: Option<usize>,
-    input: Option<u32>,
-) -> PgoConfig {
-    match pgo {
-        PgoType::Cell(cli_max_columns) => {
-            let pc_idx_count = powdr_openvm::execution_profile_from_guest(
-                &guest,
-                guest_opts.clone(),
-                stdin_from(input),
-            );
-            assert!(
-                cli_max_columns.is_none(),
-                "cli --pgo can't parse Cell(Option<usize>), input must be wrong"
-            );
-            PgoConfig::Cell(pc_idx_count, max_columns, max_block_instructions)
-        }
-        PgoType::Instruction => {
-            let pc_idx_count = powdr_openvm::execution_profile_from_guest(
-                &guest,
-                guest_opts.clone(),
-                stdin_from(input),
-            );
-            PgoConfig::Instruction(pc_idx_count)
-        }
-        PgoType::None => PgoConfig::None,
-    }
 }
 
 fn setup_tracing_with_log_level(level: Level) {
